requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.27.1
google.generativeai
gtts